Moved classes into new files in lib folder for better organisation.
