+++ lib/signup_screen.dart
@@ -1,0 +1,115 @@
+import 'package:flutter/material.dart';
+import 'widgets/widgets.dart';
+
+class SignUpScreen extends StatefulWidget {
+  const SignUpScreen({super.key});
+
+  @override
+  State<SignUpScreen> createState() => _SignUpScreenState();
+}
+
+
+class _SignUpScreenState extends State<SignUpScreen> {
+  String firstName = '';
+  String surname = '';
+  String username = '';
+  String password = '';
+  String repeatPassword = '';
+  String errorMessage = '';
+
+  void onBackButtonPressed() {
+    print('Back button pressed.');
+    Navigator.pop(context);
+  }
+
+  void signUp() {
+    print('nothing so far');
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: const Color(0xFFACD4AE),
+      appBar: AppBar(
+        backgroundColor: const Color(0xFFACD4AE),
+        leading: IconButton(
+          icon: Icon(Icons.arrow_back, color: const Color(0xFF326335),),
+          onPressed: onBackButtonPressed,
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Sign Up',
+              style: Theme.of(context).textTheme.displayLarge,
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: TextField(
+                onChanged: (value) => username = value,
+                decoration: InputDecoration(
+                  label: Text('First Name', style: Theme.of(context).textTheme.bodyMedium),
+                  border: OutlineInputBorder(),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: TextField(
+                onChanged: (value) => username = value,
+                decoration: InputDecoration(
+                  label: Text('Surname', style: Theme.of(context).textTheme.bodyMedium),
+                  border: OutlineInputBorder(),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: TextField(
+                onChanged: (value) => username = value,
+                decoration: InputDecoration(
+                  label: Text('Username', style: Theme.of(context).textTheme.bodyMedium),
+                  border: OutlineInputBorder(),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: TextField(
+                onChanged: (value) => password = value,
+                decoration: InputDecoration(
+                  label: Text('Password', style: Theme.of(context).textTheme.bodyMedium),
+                  border: OutlineInputBorder(),
+                ),
+                obscureText: true,
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: TextField(
+                onChanged: (value) => password = value,
+                decoration: InputDecoration(
+                  label: Text('Repeat Password', style: Theme.of(context).textTheme.bodyMedium),
+                  border: OutlineInputBorder(),
+                ),
+                obscureText: true,
+              ),
+            ),
+            PrimaryButton(label: 'Sign Up', onPressed: signUp),
+            if (errorMessage.isNotEmpty)
+              Padding(
+                padding: const EdgeInsets.only(top: 10),
+                child: Text(
+                  errorMessage,
+                  style: TextStyle(color: Colors.red),
+                ),
+              ),
+          ],
+        ),
+      ),
+    );
+  }
+}
